+++ node/module.ts
@@ -754,9 +754,21 @@ for (const key of Object.keys(nodeMods)) {
 nativeModulePolyfill.set(
   "module",
   createNativeModule("module", {
-    default: Module,
+    _cache: Module._cache,
+    _extensions: Module._extensions,
+    _findPath: Module._findPath,
+    _initPaths: Module._initPaths,
+    _load: Module._load,
+    _nodeModulePaths: Module._nodeModulePaths,
+    _pathCache: Module._pathCache,
+    _preloadModules: Module._preloadModules,
+    _resolveFilename: Module._resolveFilename,
+    _resolveLookupPaths: Module._resolveLookupPaths,
     builtinModules: Module.builtinModules,
     createRequire: Module.createRequire,
+    globalPaths: Module.globalPaths,
+    Module,
+    wrap: Module.wrap,
   }),
 );
 
@@ -1359,7 +1371,7 @@ export function resolveMainPath(main: string): undefined | string {
   // Note extension resolution for the main entry point can be deprecated in a
   // future major.
   // Module._findPath is monkey-patchable here.
-  let mainPath = Module._findPath(path.resolve(main), [], true);
+  const mainPath = Module._findPath(path.resolve(main), [], true);
   if (!mainPath) {
     return;
   }