+++ node/module_all.ts
@@ -10,6 +10,7 @@ import dns from "./dns.ts";
 import events from "./events.ts";
 import fs from "./fs.ts";
 import fsPromises from "./fs/promises.ts";
+import internalReadlineUtils from "./internal/readline/utils.js";
 import http from "./http.ts";
 import net from "./net.ts";
 import os from "./os.ts";
@@ -17,12 +18,15 @@ import path from "./path.ts";
 import perfHooks from "./perf_hooks.ts";
 import process from "./process.ts";
 import querystring from "./querystring.ts";
+import readline from "./readline.ts";
 import stream from "./stream.ts";
+import streamWeb from "./stream/web.ts";
 import stringDecoder from "./string_decoder.ts";
 import sys from "./sys.ts";
 import timers from "./timers.ts";
 import timersPromises from "./timers/promises.ts";
 import tty from "./tty.ts";
+import vm from "./vm.ts";
 import url from "./url.ts";
 import util from "./util.ts";
 
@@ -31,11 +35,9 @@ import util from "./util.ts";
 // import dgram from "./dgram.ts";
 // import http2 from "./http2.ts";
 // import https from "./https.ts";
-// import readline from "./readline.ts";
 // import repl from "./repl.ts";
 // import sys from "./sys.ts";
 // import tls from "./tls.ts";
-// import vm from "./vm.ts";
 // import workerThreads from "./worker_threads.ts";
 // import zlib from "./zlib.ts";
 
@@ -53,13 +55,16 @@ export default {
   fs,
   "fs/promises": fsPromises,
   http,
+  "internal/readline/utils": internalReadlineUtils,
   net,
   os,
   path,
   "perf_hooks": perfHooks,
   process,
   querystring,
+  readline,
   stream,
+  "stream/web": streamWeb,
   "string_decoder": stringDecoder,
   sys,
   timers,
@@ -67,5 +72,6 @@ export default {
   tty,
   url,
   util,
+  vm,
   zlib: {},
 } as Record<string, unknown>;